dapp: decode of utxoData containing 2 or 3 UTXOs
Without this patch, the dApp assumed that the utxoData has 3 UTXOs
always. That is not the case for the UTXO created via bridging of the
stakes from the Goerli network.

This is a problem because it breaks the decoding of the utxoData.

This patch solves the problem by making a temporary fix that assumes
that the utxoData can have both 2 or 3 UTXOs.
Issue-URL: https://app.clickup.com/t/383vkue