dapp: move formatNumber() outside ClaimedProgress
move the formatNumber function outside the ClaimedProgress components

Without this patch, the formatNumber function is a pure function defined
inside the ClaimedProgress component.

This is a problem because this function is redefined every time the
component is rendered.

This patch solves the problem by moving the formatNumber() function
outside ClaimedProgress

Issue-URL: https://app.clickup.com/t/2unhe59