dapp: dynamically change the staking input font size
Without this patch, the staking input font size is fixed regardless the
length of the text inside it.

This is a problem because when we continue typing, the text became larger
than the area to view it.

This patch solves the problem by shrinking the font size according to
the text length

Issue-URL: https://app.clickup.com/t/2rh81bz