dapp: fetch the ZKP market price on app load
Without this patch, the app fetches the ZKP market price only on the
load of the staking page.

This is a problem because if zAssets page was open first the total
balance in USD is not shown.

This patch solves the problem by dispatching the action to fetch the ZKP
market price on app load.

Issue-URL: https://app.clickup.com/t/389h9p4