dapp: prevent normal zAssets page on wrong network
Without this patch, the normal zAssets page is being
shown on the wrong networks (networks which have no MASP).

This is a problem because the user must be informed that
in order to deal with the zAssets list, they must be
switched on appropriate networks.

This patch solves the problem by calling chainHasPoolContract() in the services/contracts.ts
and then by showing a new component called "WrongZassetsNetwork"
inside the zAssets page if current chain has no MASP.

Issue-URL: https://app.clickup.com/t/2cdvy5z