contracts: fix solhint Cannot read property 'errorCount' of undefined
Without this patch, there are situations when the .solhintignore file
causes all inputs files to be ignored. This is a problem because the
'reports' variable ends up being an empty array, and therefore
reports[0] is undefined.

This patch solves the problem by creating a forked version of solhint
and making changes in checking that the reports array exists and has a
length greater than 0.

Pull request in solhint is created:
https://github.com/protofire/solhint/pull/351

Untill then, we will use our forked version.