contracts: fix solhint not ignoring .solhintignore specified files
Without this patch, the solhint linter did not ignore files specified in
the .solhintignore file.

This is a problem because lint-staged returns linting issues for the
files that are ignored in the repository and did not allow to commit the
changes due to git commit hook error.

This patch solves the problem by fixing the paths of the files specified
in the solhint command. More specifically, by changing full the paths to
be relative to the contracts workspace paths.

Issue-URL: https://app.clickup.com/t/38ukmjf