dapp: some SVG icons disappear on browsers
without this patch, when we use SVG icons with filter property, they
don't appear on Firefox and Safari, while they work well with Chrome.

this is a problem because it breaks the UI with some major browsers.

this patch solved the problem by using PNG files in such cases instead
of using SVG files.

Issue-URL: https://app.clickup.com/t/2rrcrhv