dapp: update zAssets table width
Without this patch, regarding the dateTime format on user machine, the
redeem button in zAssets rows may needs three line to view the "unlocked
until" title.

This is a problem because it breaks the UI because of its need to some
extra spaces.

This patch solves the problem by increasing the table width a little bit
to fit the maximum dateTime size.

Issue-URL: https://app.clickup.com/t/3acuvu1