dapp: fix update of the exit time on network switch on ZAssets page
Without this patch, when a user switches to any network from Mumbai,
goes to the zAssets page and switches back to Mumbai, he finds missing
dates on the redeem buttons.

This patch solves the problem by dispatching the action to update the
exit time on the zAssets page component loading.

Issue-URL: https://app.clickup.com/t/3ar8fax