dapp: refactor stake() to use keys and web3WalletLastAction states
Without this patch, the root reading and spending keys are generated in
the stake().

This is a problem because it violates layer boundary abstraction.

This patch solves the problem by generating the root reading and
spending keys before calling stake() and passing them as a parameter.
Also, it partially addresses layer boundary violation of notifyError().

Issue-URL: https://app.clickup.com/t/3967773