contracts: solhint failure on checking Proxy.sol fixed
This commit fixes a failure of `solhint` complaining on "inline assembly" and
"low level calls" in the `common/proxy/Proxy.sol`.

1. Solhint "disable" directive in the Proxy.sol updated to disable warning for
"inline assembly" and "low level calls" (the Proxy.sol is a 3rd party contract
and its code must remains "as is", with the assembly code and low level calls).

2. Solhint version updated to 3.3.7.

3. `.solhintignore` removed since this file make `solhint`, due to a known bug
in the latest, breaks CI script (invoked via `.lintstagedrc.mjs`):
with `.solhintignore` given to `solhint`, if 'lint-stage' selects no solidity
files for linting, `solhints` reverts with non-zero exit code because of the
"Cannot read property 'errorCount' of undefined" bug; this bug is dormant if
`.solhintignore` is not used.