dapp: update design of PrivateBalance component
Without this patch, we have not synced our
PrivateBalance component in zAssets page
with latest design changes in version 0.5.

This is a problem because we need to update our
frontend styles according to latest changes.

This patch solves the problem by doing some
psd-to-html job according to the figma link in
the task description.

Issue-URL: https://app.clickup.com/t/38unhf4