dapp: fetch rewards with max retries and exponential backoff delay
Without this patch,  our staking page does not
update the UTXO status automatically after the
staking trx is successfully done.

This is a problem because user expects to see
the rewards get updated after staking.

This patch solves the problem by fetching new
UTXO from the subgraph and updating the UTXO status
using the signatures available after the staking.
We did this by updating getAdvancedStakesRewards action in the
advancedStakesRewards.ts slice, which sequentially fetches,
then gets from the state the updated rewards, and
then finally refreshes the statuses.

Issue-URL: https://app.clickup.com/t/39nzgxj