dapp: hook up refresh rewards button in staking page
Without this patch, we lack a handy functionality
in our BalanceCard component to refresh the user
staking rewards.

This is a problem because users need to be able to
manually refresh the rewards.

This patch solves the problem by creating a separate
hook in web3.ts called useRefreshAdvancedStakesRewards
which calls getAdvancedStakesRewards() function from
advancedStakesRewards slice in order to refresh the
rewards.

Issue-URL: https://app.clickup.com/t/3c06548