dapp: fix MetaMask gets disconnected on the page reloading
Without this patch, whenever a user reloads the page the wallet is
disconnected.

This is a problem because it results in a bad user experience.

This patch solves the problem by saving in the Persist Redux Store the
boolean value isWalletConnected. Then, on the page reload the value is
retrieved in the WalletHeader component and if it is true, the wallet is
re-connected.

Issue-URL: https://app.clickup.com/t/33t4w1c