dapp: Remove CSS variables only used once
Without this patch, we have a tech dept
which is about having unnecessary custom
variables created in scss files for the
values which only used onece.

This is a problem because it's a clear
violation of DRY principle.

This patch solves the problem by removing
all those variables which has only one
implementation case in same scss file.

Issue-URL: https://app.clickup.com/t/2c0ecyq