dapp: staking 2 times in a row
Without this patch, we can click the stake button as much as we want.

This is a problem because it trigger more than one transaction in a row

for the same staking process

This patch solves the problem by disabling the staking button when there

is a staking process in progress.

Issue-URL: https://app.clickup.com/t/2ky8zfh