contracts: Modify `rewards:limit:add` task Without this patch, the `rewards:limit:add` is not compatible with the new version of AdvancedStakeRewardController.
This is a problem because the AdvancedStakeRewardController now has 2
extra functions for setting/updating the rewards limit. One function
handles the ZKP and PRP rewards
limit and another one handles NFT reward limit.

This patch solves the problem by modifying the `rewards:limit:add` to
be compatible with the new version of the AdvancedStakeRewardController.

The AdvancedStakeRewardController
updates the ZKP and PRP rewards limit based on its ZKP and PRP balances.
But since it does not hold any NFT, so the desired NFT Rewards Limit
should be defined for contract. then the contract will be able to mint
NFT up to that amount.

example to execute the task:
```
yarn hardhat rewards:limit:add --nft-limit 10  --zkp-prp-limits true --network mumbai
```

Note that the `--nft-limit` and `--zkp-prp-limits`
are optional task parameters.

Issue-URL: https://app.clickup.com/t/2qmjevy