dapp: move errors and notification files outside servics layer folder
Without this patch, errors and notification files are located in services
folder which belongs to service layer.

This is a problem because both files use UI element which make it unacceptable
to have them in services layer.

This patch solves the problem by moving both files into common folder
under components folder.

Issue-URL:  https://app.clickup.com/t/2f72g27