dapp: delete connect wallet link
Without this patch, there is a broken connect wallet link at the top of the
balance card.

This is a problem because the link is broken, and it is a UI overwhelming
since we have the connected button in two places.

This patch solves the problem by deleting this extra link.

Issue-URL: https://app.clickup.com/t/3c7c4az