dapp:mobile-panels-reordering
Without this patch, we have a tiny issue in how staking page is
displayed in terms of the order of the components in mobile screen
sizes.

This is a problem because we need to display the APY banner component on
the top of page.

This patch solves the problem by giving relative position to
.advanced-staking-rewards & .balance-card classes in media query.

Issue-URL: https://app.clickup.com/t/238r8dx