dapp: fix spinner shows up on all "Redeem" buttons
Without this patch, after calling the redeem function, all the redeem
buttons in zAssets page get spinning icon inside of them.

This is a problem because we only need to have the selected button get
spinning icon.

This patch solves the problem by adding a new local state in
RedeemRewards component called selectedRewardId and conditionally showing
spinning icon if rewrads id is equal to selectedRewardId.

Issue-URL: https://app.clickup.com/t/3e6dxkp