dapp: add dev file of the built tres in .gitignore
Redeem functionality for the dapp requires the tres file of the
NewCommitments events. This file needs to be updated on every new
generation of the UTXO and continuously changing. In production,
however, AWS will serve this file from the S3 bucket. For development
purposes, the file is added to the .gitignore file.