dapp: add redux slice for dismissed notifications
Without this patch, the MesseageAlert component shows up every time the
user reopens the zAssets page.

This is a problem because the user may have already dismissed the
notification but it keeps bothering him.

This patch solves the problem by creating a new slice called
acknowledgedNotifications which persists the notif state for every page
that uses MessageAlert component

Issue-URL: https://app.clickup.com/t/3cjrwga