dapp: fix geo location checking bug
without this patch, the user is prohibited if the geolocation APIs are not
reachable.

this is a problem, because if we can't detect if the user is from prohibited
jurisdiction then we mustn't prohibit him.

this patch solved the problem by updating the logic to allow the user to
use the app if we couldn't detect his/her geolocation data.

Issue-URL: https://app.clickup.com/t/2yywz4a