dapp: feat hook up 2-step exit function
Without this patch, the user is not able to perform redeeming of the
available rewards. In the recent version of the smart contract, the
rewards are redeemed in 2 steps to avoid the reentrancy attack. The
first step is to register the commitment to "exit" and the second is to
call the `exit` function.

This patch solves the problem by adding functionality in dApp to perform
the 2-step exit function. First, registerCommitToExit() function is
created to register the commitment in the smart contract. After the
commitment is registered, the user can perform the exit() function when
the delay timer passes. The patch updates UI elements to allow the
user to perform registerCommitToExit() call. Also, the patch fixes the
bug in the openNotification() function. The function was not working
properly because the 'details' field in the DetailedError was an object.

Issue-URL: https://app.clickup.com/t/3dn9vk6