dapp: move chainHasStakesReporter debug to service layer
Without this patch, we have not followed the separation of concerns in
our dapp for chainHasStakesReporter debug message usage.

This is a problem because this debug message does not belong to
totalUnclaimedClassicRewards slice.

This patch solves the problem by deleting the debug message from
totalUnclaimedClassicRewards and move it inside getStakesAndRewards()
function instead.

Issue-URL: https://app.clickup.com/t/2gwn4fn