dapp: feat change of encryption to AES-128-CBC
Without this patch, the encryption is AES-256-CBC. The new design of the
protocol specifies AES-128-CBC.

This patch solves the problem by changing the encryption to AES-128-CBC.
Both cipher key and IV are generated from a shared packed ECDH key split
in half. The first half of the packed key is used as the cipher key
because it has more entropy, and the second half is used as IV. The
cipher key length and IV length are both 16 bytes.

Issue-URL: https://app.clickup.com/t/2kr934u