dapp: restrict user to add maximum 18 digits after period
Without this patch, if the user write 19 digits after the period in the
staking input field, the staked amount will be null.

This is a problem because the user needs to enter the value from the beginning.

This patch solves the problem by alerting the user and restrict him to be
able to add by maximum 18 digits after the period.

Issue-URL: https://app.clickup.com/t/2yyr71a