dapp: update welcome page
without this patch, the welcome screen doesn't fit with Figma design

this is a problem, because all pages must reflect exactly Figma design.

this patch solves the problem by updating the page to fit with design.

this patch also deleted some code that is commented using // and there
is no needs for it anymore.

Issue-URL: https://app.clickup.com/t/390fy2v