dapp: feat storybook support
Without this patch, we will not be able to have a storybook and not be
able to have a visual check of our components during the implementation.

This is a problem because we had a lack of tests.

This patch solves the problem of non-having visual tests during and
after the implementation process, also during this implementation, I
changed the webpack version and updated stuff according to this; I had
to move webpack from ts to be in the js file for later will leave
further investigation about moving it back to ts.

Issue-URL: https://app.clickup.com/t/3dfjb5x