dapp: disable Merkle proof verification via calling smart contract
Without this patch, the dApp would call the smart contract to verify the
Merkle path of the exit. This is not possible anymore since the smart
contract does not have the Merkle tree verification code. The code was
present in the smart contract for developing purposes only and removed.

This is a problem because the dApp would not be able to exit and throw
an error.

This patch solves the problem by removing the verification call to the
smart contract. The smart contracts will still verify the Merkle path
during the exit call().

Issue-URL: https://app.clickup.com/t/3ggr514