dapp: hardcoded PRP rewards
Due to the recent changes in the smart contract, the PRP UTXOs are not
being generated upon staking.

This patch solves the problem by hardcoding the PRP rewards. Due to the
specs, only the first 2000 stakes will get 2000 PRPs, and the rest will
get 0 PRPs. The hardcoded values should be removed upon deployment of
version 1.0 of the protocol. MASP doc:
https://docs.google.com/document/d/1BTWHstTgNKcapOe0PLQR41vbC0aEDYmbBenfzTq8TVs

Issue-URL: https://app.clickup.com/t/3gw3pzu