dapp: feat update encryption scheme of the smart contracts
Without this patch, the encryption message has a size of 48 bytes.
However, in the most recent specs of the MASP, the encryption message
has a size of 32 bytes.

This patch updates the encryption scheme of the smart contracts to have
a message size of 32 bytes. This is done by removing the padding of the
message. Moreover, IV is removed from the packed message as the IV is
generated from the packed ephemeral public key. The patch also updated
the test cases to reflect the changes.

Issue-URL: https://app.clickup.com/t/3yaw68b