dapp: disable PRP grantor smart contract in dApp
Without this patch, dApp fetched the data about the remaining PRP
rewards from the PRP grantor smart contracts. In the most recent specs
of the smart contracts, the advanced stakes do not generate PRP rewards
as UTXO in the pool.

This patch disables fetching the data about PRP rewards from the smart
contracts.

Issue-URL: https://app.clickup.com/t/3yaw68b