dapp: feat early redeeming with new smart contracts
Without this patch, the redeeming of the dapp is not possible. In the
new version of the smart contracts, 225 bytes is the size of the UTXO
data containing 1 zZKP UTXO, 1 NFT UTXO generated during advanced stake.
First byte is reserved for the msg version number. Next 96 bytes after
the Message type are: 64 bytes are packed message with UTXO secrets, 32
bytes for the zAssetsID. The last 128 bytes are those of the NFT data
(ignored for now).

This patch solves the problem by updating the exit function that works
with new smart contracts. First, the patch updates the method to decode
the UTXO data that contains a new message formatting. Second, decoded
zAssetID from the UTXO data is compared to the expected zAssetID from
Registry smart contract. Lastly, the latest version of the advanced
stake generates only two spending keys: for zZKP and NFT. Finally, the
patch updates the calculation of the advanced stake PRP rewards based on
the leaf ID. Also, it updates the the subgraph querry and the ABI of the
AdvancedStakeRewardController.

Issue-URL: https://app.clickup.com/t/3yaw68b