contracts: add a task to test bridge on forked network Without this patch, there is no easy way to test the bridge on a forked network.
This patch solves the problem by adding the
`staking:bridge:local` task.

The message passes through these contracts
`msgSender@goerli => fxRoot@goerli => fxChild@mumbai => msgRelayer@mumbai`

The fxRoot and fxChild contracts have been implemented/deployed by Polygon
team. they handles all of the brdige requests that come from any contract,
hence it's very hard wathing them to dubug our own contracts.

We are usually able to send a message from our msgSender contract to
polygon's fxRoot without any problem, but most of the times, the fxChild
cannot send the retrived message to msgRelayer. it can be cause of an error in the our polygon contracts:
`msgRelayer`, `rewardMaster`, `advancedStakeRewardController`, `pool`, etc.
So the best way to debug the Bridge is to impersonate the fxChild and
send a message to msgRelayer on behalf of it.
The `staking:bridge:local` has been implemented to achieve that.
This task sends a tx on behalf of `fxChild` to the
`advancedStakeActionMsgRelayer` contract to
execute its `processMessageFromRoot` function.

Issue-URL: https://app.clickup.com/t/2xjxen3