dapp: fix footer position
Without this patch, our footer has mush distance from the bottom of the screen.

This is a problem because it has not a good user experience.

This patch solves the problem by revising the min-height body-container
& footer-container inside MainPageWrapper/styles.scss file. Also, it
adds importing the styled-scroll-bar @mixin into the styles.scss which
has been already created in mixins.scss

Issue-URL: https://app.clickup.com/t/30xq8rc