dapp: enable clearing stakingAmount in all cases
Without this patch, when switching the account to another one with insufficient
balance, we can't clear the staking amount input field if it already contains some value.

This is a problem because we need to clear it to add the amount that is
less or equal to the token balance value.

This patch solves the problem by enabling the cleaning of the input field regardless
any conditions.

Issue-URL: https://app.clickup.com/t/30xqkrg