dapp: fix failing test for ClaimedProgress component
Without this patch, we have failing test for ClaimedProgress.

This is a problem because the component implementation is changed so now the
test is out-dated so it must be revised according to the new changes.

This patch solves the problem by changing '10k out of 10m' to '10k / 10m' in
ClaimedProgress.test.tsx file.

Issue-URL: https://app.clickup.com/t/31atnf3