contracts: avoid redeploying if contract is in env
Without this patch, the contracts are redeployed regardless of the env
vaiables. Plus, the env variables are very
general and have not been seperated
by network, while some of them can be exists on more than one
network.

This patch solves the problem by adding the `reuseEnvAddress` that skips
deployment if the env address for the contract has been provided.
This patch also adds a `getContractAddress` into `deploymentHelpers.ts`
to retrieve contract address from env variables or deployment artifacts.