dapp: check zero padding of the secret random upon deciphering
Without this patch, we are not checking if the secret random is padded
with zeros. All secret randoms should be padded with five zeros in
binary representation because we use babyjubjub.suborder() as the
modulus. This patch adds a check to ensure that the deciphered random
values are padded with zeros.

Issue-URL: https://app.clickup.com/t/3k6p09r