Slither vulnerability - shadowing-state fix
Without this patch, Slither will complain about shadowing-state
vulnerability.

This is a problem because the variable __gap is for upgrades, it is not
shadowing any variable through inheritance, so it is a false positive.

This patch solves the problem by silencing the vulnerability on the line
of code which is causing this vulnerability.

Issue-URL: https://app.clickup.com/t/2wv0tj1