Slither vulnerability - unchecked-transfer
Without this patch, Slither will complain about unchecked-transfer
vulnerability.

This is a problem because we are calling the .transfer function to
ZKPToken contract, so safe transfer is not mandatory. So it is a false
positive.

This patch solves the problem by silencing the vulnerability on the line
of code which is causing this vulnerability.

Issue-URL: https://app.clickup.com/t/2wv0wt5