dapp: show zero in APR field when wallet not connected
Without this patch, we show same thing in APR section on top
of staking page both when wallet is connected and when not.

This is a problem because we must show different UI when
wallet is not connected.

This patch solves the problem by showing zero in APR section
if wallet not connected.

Issue-URL: https://app.clickup.com/t/325y630