dapp: fix bug in registering the exit commitment
Without this patch, if an error occurs during registering of the
commitment happened, UTXO status is not updated. This became a problem
when UTXO was spent; thus, the registerCommitToExit() function returns
an error. As a result, UTXO has the UNSPENT status and keeps persisting
on the account list as unspent UTXO.

This patch solves the problem by returning the status from
registerCommitToExit() function call. In case of an error, the status is
updated.

Issue-URL: https://app.clickup.com/t/3qzmw4h