dapp: fix background color of unstake tab list
Without this patch, we have lost background color of
unstake tab list due to the new changes in global styles
in other related commits.

This is a problem because the background color of other
places must not get affected when applying global style
refactoring.

This patch solves the problem by reseting back the previous styles
locally to the related scss file.

Issue-URL: https://app.clickup.com/t/3kqnvtu