dapp: fix connect button text in faucet page
Without this patch, we have a tiny disalignment of text
for ConnectButton when wallet is disconnected in faucet page.

This is a problem because we must keep texts aligned at center.

This patch solves the problem by removing the wrapper box
and revise the name of data-testId in the related test file
to avoid breaking the component test.

Issue-URL: https://app.clickup.com/t/34hth1k