dapp: splitting exit function
Without this patch, the exit function is too long.

This is a problem because it is  hard to read and maintain and may cause
DRY violation.

This patch solves the problem by splitting the exit function into smaller
functions with specific functionality for each one through:

- destructuring the code that checks and returns if zAsset is not registered
into a separate function.

- hiding the implementation of the Poseidon function to calculate the commitmentHex

- add a function  getMerklePath which returns either error merkle path

Issue-URL: https://app.clickup.com/t/2h83xaj