dapp: capture try-catch err with Sentry
Without this patch, we have no way to record users catched error in a place we can review.

This is a problem because we need to review those errors to find a way to solve them and rise the quality of the code.

This patch solves the problem by using Sentry captureException function.

Issue-URL: https://app.clickup.com/t/3ufvvf5

Issue-URL: https://app.clickup.com/t/3ufvvvx