dapp: feat use all subgraph accounts until a successful request
Without this patch, if the subgraph fails to load and returns an error
for any reason, the dapp will not be able to load and display the data.
With this patch, the dapp will try to load the data from another
subgraph account recursively until it succeeds or all accounts are
tried.