dapp: refactoring dropdown list component
Without this patch, we are using FormControl and Select components to
produce dropdown list component.

This is a problem because it is not the best way that material ui
documentation use and suggest.

This patch solves the problem by using Menu and MenuItem.

Issue-URL: https://app.clickup.com/t/37k8kya

This patch also solved another bug which is addressed in this task:

Issue-URL: https://app.clickup.com/t/344vqtz