dapp: update unstake tab content
Without this patch, we miss a new column in
unstake tab which is called Expected.

This is a problem because we need to include
this new column to show the user the expected
prp related to that stake row.

This patch solves the problem by adding Expected
column in UnstakeTab filled with related data.

Issue-URL: https://app.clickup.com/t/34g714c