dapp: change unstaking button color when active
Without this patch, the colour of the unstake button is grey.

This is a problem because it does not match other active colours
(for example, blue "redeem button")

This patch solves the problem by changing the unstake button's colour
to blue as the redeem button has.

Issue-URL: https://app.clickup.com/t/3werm78

this patch also fix the bug addressed in task (#3wernqw) to keep Unstake row to
fit the line without scrolling

Issue-URL: https://app.clickup.com/t/3wernqw