dapp: fix reject transaction submission with invalid EIP712 signature
Without this patch, the transaction is submitted even if the EIP712
signature is invalid. With the invalid transaction, the user loses gas
upon submission, as the smart contract reverts it in any case. This
patch solves the problem by checking the signature before submitting the
transaction and showing an error if the signature is invalid. This patch
does not fix the bug that produces an invalid signature; it only ensures
that the transaction is not submitted with an invalid signature.

Issue-URL: https://app.clickup.com/t/3wmap76