contracts: add minter for PNFT
Without this patch, The nft token can be minted only by owner.

This is a problem because, the owner is changed immediately after
deploying the PNFT contract, however, the baseURI and contractURL
still need to be updated by the owner in the PNFT contract.
That can be a problem if the owner is a contract, i.e, rewardController.

This patch solves the problem by adding a minter role that can only mint
tokens. This value is updated immediately after deploying the contract.
The owner is also immediately changed to multisig, but it guarantees that
in the future the token and contract URIs can be set through a DAO
proposal.

Some unit tests have also been added for `setContractURI` and
`setBaseTokenURI` functions.

Issue-URL: https://app.clickup.com/t/320ugjv