contracts: fix the bug when fill up local envs
Without this patch, the constant env variable for testnet/main networks
are replaced with constant env variable for localhost network.

This is a problem because of a bug inside the `fulfillLocalAddress()`.
Since some env variables for localhost are filled up by defult
(i.e, `OPENSEA_ERC721_PROXY_REGISTRY_LOCALHOST` ), the
`fulfillLocalAddress()` sees this value and add it to
`OPENSEA_ERC721_PROXY_REGISTRY` regardless of the network name.
however, it should be done only if the network name is one of
'LOCALHOST', 'PCHAIN', 'HARDHAT' values.