dapp: update APR field with env value
Without this patch, we lack the getPrpAPY()
function to get APY PRP.

This is a problem because we need it to be
fetched from env file as hardcoded value.

This patch solves the problem by adding APY_PRP
as variable in env files and also by creating
a new function in rewards.ts file called
getPrpAPY() which returns the variable.

Issue-URL: https://app.clickup.com/t/37608gb