dapp: feat calculate unrealized privacy rewards
We have two types of rewards (APRs) in advanced staking. The first one
is applied to the staked ZKP amount to calculate zZKP rewarded. It is
fixed and is already voted upon and implemented. The second APR (subject
of this patch) is applied to the amount of zZKP remaining in the MASP to
calculate unrealized (accrued) PRP rewards. The actual PRP amount will
be calculated when a user spends zZKP (after the CoreV1 launch). APR #2
has yet to be voted for and implemented.

To estimate unrealized rewards (the subject of this patch), we are
applying the following formula:

unrealized_PRP_reward =
zZKP_amount * zkp2prp_rate * APR_2 *(time_now - time_staked) / 365 days

where,
- prp2zkp_rate = 1/10 - The conversion rate of 10 PRPs for 1 ZKP is
  supposed to be an initial value set to the AMM. It is already voted
  for. But the AMM still needs to be implemented.
- APR_2 = 10% - 10% per annum will be proposed for voting. But it may be
  another value.

Issue-URL: https://app.clickup.com/t/375y072