dapp: split the file dapp/src/lib/constants.ts
Without this patch, we have both number constants and contracts-confirmations
constants in one file.

This is a problem because it is more readable and cleaner to have each
constants group in its separate file.

This patch solves the problem by spliting the file into two files inside
constants folder directly under src folder.

Issue-URL: https://app.clickup.com/t/3wu3d6y