dapp: fix max zZKP reward limit based on exceeding reward in the pool
Without this patch, we allow user staking in the UI if the rewards are
more significant than we have in the pool. This is a problem because the
smart contract will revert the transaction with a cryptic error. This
patch fixes the problem by checking available rewards in the smart
contract and limiting the stake if the rewards are more significant than
we have in the pool.