dapp: simplify disabled logic of redeem button
Without this patch, we have a complex logic for the
redeem buttons apeared in zAssets page list.

This is a problem because we must keep code simple
and more readable.

This patch solves the problem by splitting the logic
into a few variables assigned to a wrapper variable
called isRedemptionPossible.

Issue-URL: https://app.clickup.com/t/3wg1xpk