dapp: disable all redeem buttons if wallet in progress
Without this patch, we have a bug when another action
is in progress and user clicks on a different redeem
button opening the first redeem modal and after he
clicks the check box, the modal button is still disabled
because another action is already in progress.

This is a problem because the user should not be able
to make another action if other wallet action is
already under progress.

This patch solves the problem by adding an extra
condition for the redeem buttons disabled property
to prevent more actions while wallet is in progress.

Issue-URL: https://app.clickup.com/t/37hfxky