dapp: fix showing final redeem button
Without this patch, we need to close the redeem dialog and click the
redeem zZKP button again to open it so we can see and click the final
redeem button.

This is a problem because it is not user-friendly solution and add another
extra step to the redemption process.

This patch solves the problem by showing the button automatically when
the period is passed.

Issue-URL: https://app.clickup.com/t/37hfeuc