dapp: feat add isWalletUpdating redux state
Without this patch, the dApp is missing a global state when the wallet
progresses with some updates or calculations in the background.

This is a problem because, in some cases, the user needs to be notified
of actions in the background.  These actions may update the state of the
wallet, but the user is not informed about them.

This patch solves the problem by adding a new slice in redux/UI reducer
called isWalletUpdating.  Also, the patch adds the spinner (and tooltip)
in the Balance Card if isWalletUpdating is in progress.

Issue-URL: https://app.clickup.com/t/3x5m06e