dapp: unify balance styling
Without this patch, the balance in unstaked balance is styled in a way
differ from the other balances in the staking page/balance card.

This is a problem because we need to unify the styling of all balances.

This patch solves the problem by creating a component that return styled
balance and use it.

Issue-URL: https://app.clickup.com/t/37wvgvt