fix(dapp): app freezes when render 100+ stakes
Without this patch, the app will freeze when trying to render 100+ stakes.

This is a problem because, the dapp is not usable at this state.

This patch solves the problem by removing the `backdrop-filter: blur(36px)` CSS
effects which caused a performance drop when trying to render a lot of
components (stakes) at once, then fix the broken layout introduced by removing
the `blur` CSS property

Issue-URL: https://app.clickup.com/t/382pzx4