feat(dapp): make pagination component responsive
Without this patch, we have not a responsive
pagination component which we already created.

This is a problem because it must be responsive.

This patch solves the problem creating a new varibale
constant called MAX_PAGE_LIMIT_MOBILE inside
src/constants/pagination.ts which will be conditionally
used by using useScreenSize() hook.

Issue-URL: https://app.clickup.com/t/861m3he89