refactor(dapp): Remove routes code duplication
Without this patch, we will have duplicate links in the dApp header and footer.

This is a problem because haveing duplicated code makes it hard to maintain or
update it in the future.

This patch solves the problem by moving all the routing/links logic into its own
file. This makes it easier to update the label and URL of the link. Control the
dApp routes from single `ENUM` which makes it easier to update as well

Issue-URL: https://app.clickup.com/t/3xpkudn