feat(dapp): create a resuable hook for pagination
Without this patch, we have some DRY cases in
our dapp, where we repeated some pagination
functions and states.

This is a problem because we better abide by
DRY rule.

This patch solves the problem by creating a new
hook called usePagination which does all
the implemention logic and returns paginated
data to the component in which it is called.

Issue-URL: https://app.clickup.com/t/861m3he89