fix(dapp): generation of subgraph query
Without this patch, the generation of the subgraph query fails. It fails
because getCommitmentsQuery() has the check that the leaf range does not
exceed the limit of 1000 (the subgraph service imposes this limit).
However, each triad entity has four leaves, and this was not taken into
account. As a result, the function getCommitmentsQuery() throws the
error. This patch solves the problem by dividing the leaf id range by 4.