fix(dapp): zZkp reward calculation at the end of the program
Without this patch, the rewards are miscalculated at the end of the
program. More specifically, the reward calculation assumed that end of
the advanced staking program and the reward end period are the same time
stamp. Unfortunately, this is not the case.

This patch fixes the problem, assuming that that lockedTill of the stake
is always less than rewardEnd parameter specified in
AdvancedStakeRewardController. For details, see comparisons of
lockedTill and _rewardParams.endTime in _computeZkpReward() of
AdvancedStakeRewardController contract.