feat(dapp): update wallet balance upon clicking refresh icon
Without this patch, we do not update the wallet
balance of user on the top of balance card after
clicking refresh icon.

This is a problem because user expect all fields
of balance card to get updated upon clicking
the refresh icon.

This patch solves the problem by dispatching the
getChainBalance in RefreshTokenBalanceAndRewards
function inside the UnstakedBalance component.

Issue-URL: https://app.clickup.com/t/861m4j8cx