fix(dapp): refresh UTXO on supported chains only
Without this patch, users will see error notifications whenever they switch to a
a chain that doesn't have a pool contract yet.

This is a problem because it is misleading and doesn't make any sense.

This patch solves the problem by checking if the chain has pool contract before
start refreshing UTXOs.

Issue-URL: https://app.clickup.com/t/861m4u65c