fix(dapp): Get exit time on supported chains only
Without this patch, the dApp will make unnecessary calls to the contract to get the
exit time on a chain that doesn't have a pool contract in the first place which
will be shown as error in the console.

This is a problem because this call is known that it will fail so there is no point in doing it.

This patch solves the problem by checking if the current chain has a pool contract or not.

Issue-URL: https://app.clickup.com/t/861m4u65c