fix(dapp): StyledBalance decimals default value
Without this patch, we have not set a default
value for decimal part in StyledBalance
component.

This is a problem because when wallet is not
connected we have no decimals rendered.

This patch solves the problem by adding a double
zero as deafult value in StyledBalance component
for the decimal part.

Issue-URL: https://app.clickup.com/t/861m4vxj2