refactor(dapp): improve error handling
Without this patch, we will have problems dealing with a different types of
errors like `Error` or `RPC/MM` or `DetailedError`.

This is a problem because we will have a hard time checking for which error type
are we dealing with. Also, we might lose the original error in nested functions
due to type conversion.

This patch solves the problem by creating an error wrapper around all these
error types with the ability to convert from one type to another but keep the
original state.

Issue-URL: https://app.clickup.com/t/86696a4ng