fix(dapp): get rewards if subgraph account present
Without this patch, the dapp will try to fetch the rewards from the subgraph
even if it doesn't have a subgraph account in the `.env` file.

This is a problem because this causes the dapp to show a spinner that indicates
it currently updating the balance but in reality, there is no point in waiting.

that This patch solves the problem by allowing the dapp to refresh the wallets
only if it has a subgraph account in the `.env` file

Issue-URL: https://app.clickup.com/t/861m5z5rk