feat(dapp): show console warning for users
Without this patch, regular users might be tricked into pasting malicious code
into the console.

This is a problem because this opens a door for many variabilities and
information leakage or event send TXs from the user's wallet!

This patch solves the problem by showing a big console message that warns the
user of such a case.

Issue-URL: https://app.clickup.com/t/866987jc0